+++ src/main/cpp/Main.cpp
@@ -1,9 +1,4 @@
-/*
-  Dont edit this file
-  Go to Robot.cpp
-*/
-
-#include "Robot.h"
+#include "Robot.hpp"
 #include "RobotStartup.h"
 
 WOMBAT_ROBOT_MAIN(Robot)
